#18 cart view

+++ src/main/java/eus/solaris/solaris/dto/ShopCartOutputDto.java
@@ -10,6 +10,6 @@ public class ShopCartOutputDto {
     
     List<ProductDto> products;
 
-    Long totalPrice;
+    Double totalPrice;
 
 }
